backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
numpy
openai
langchain
email-validator
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-multipart
python-dotenv